location wasn't as bas as I though. Implemented, still need to pass refs and figure out how to get both daily and current weather at same time

+++ app/src/main/java/net/rmoreno/weatherapp/WeatherInteractor.java
@@ -5,6 +5,9 @@ import com.squareup.okhttp.Callback;
 import com.squareup.okhttp.Request;
 import com.squareup.okhttp.Response;
 
+import net.rmoreno.weatherapp.presenters.Presenter;
+import net.rmoreno.weatherapp.repositories.WeatherRepository;
+
 import java.io.IOException;
 
 public class WeatherInteractor {
@@ -15,7 +18,7 @@ public class WeatherInteractor {
         this.repo = repo;
     }
 
-    public void getWeatherData(float lat, float lng, final Presenter.WeatherCallback weatherCallback) {
+    public void getWeatherData(double lat, double lng, final Presenter.WeatherCallback weatherCallback) {
         repo.getCurrentWeather(lat, lng, new Callback() {
             @Override
             public void onFailure(Request request, IOException e) {
